pyproject.toml
[tool.poetry]
name = "template-django"
version = "0.1.0"
description = ""
authors = ["Paweł Perenc <57562561+perqu@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.1"
django-rest-knox = "^4.2.0"
celery = "^5.3.6"
redis = "^5.0.3"
flower = "^2.0.1"
psutil = "^5.9.8"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
